Added saving input maps to file

+++ src/input.h
@@ -422,7 +422,7 @@ void input_mouse_mode_set(enum Mouse_Mode mode);
 int  input_mouse_mode_get(void);
 void input_update(void);
 int  input_map_state_get(const char* map_name, int state);
-void input_map_create(const char* name, struct Key_Combination* keys, size_t num_keys);
+void input_map_create(const char* name, struct Key_Combination* keys, int num_keys);
 int  input_map_keys_set(const char* name, struct Key_Combination* keys, int num_keys);
 int  input_map_remove(const char* name);
 int  input_map_name_set(const char* name, const char* new_name);
